requirements.txt
PyYAML==5.3.1
python-gitlab==1.15.0
slackclient==2.9.3
